requirements.txt
# Environment
python-dotenv==1.2.1


# GCP
google-cloud-bigquery==3.39.0


# Python src
## reader
PyYAML==6.0.3

## api
gspread==6.2.1

## general
pandas==2.3.3
db-dtypes==1.5.0